requirements.txt
ftfy>=6.0
openai>=1.0
python-dotenv
tqdm
